ivp_surface_manager: Fix downcast size_t -> int warning
Port of e4b963c from Source-Authors/source-physics.

+++ ivp_surface_manager/ivp_gridbuild_array.cxx
@@ -913,7 +913,7 @@ IVP_Compact_Grid *IVP_GridBuilder_Array::compile_ledges_into_compact_grid(const 
     cg->inv_grid_size = 1.0f / gp->grid_field_size;
 
     // reserve space for the grid elements
-    cg->offset_grid_elements = ((char *)(&cg->offset_compact_ledge_array[0])) - ((char *)cg) + sizeof(int) * ledges->len();  // skip the ledge offsets
+    cg->offset_grid_elements = ((char *)(&cg->offset_compact_ledge_array[0])) - ((char *)cg) + static_cast<int>(sizeof(int)) * ledges->len();  // skip the ledge offsets
 
     // TL: align offset to next 16 Byte Alignment (I didn't write the GridBuilder, so I have no idea wether the memory will be enough, lets hope ... - the SUN-Debugger will find it)
     int aligned_offset = cg->offset_grid_elements + 15;
